cli: retire `h5i man`, and publish the page instead
Printing a man page was a strange thing for the binary to do. It put a roff
generator and `clap_mangen` in every install to serve a file the project
already publishes a website for, and it was the only subcommand whose output
nobody read on a terminal. The page is build output. The site can hand it over:

    curl -fsSL https://h5i.dev/man/man1/h5i.1 -o ~/.local/share/man/man1/h5i.1

Removing it needed the crate split the CLI never had. `src/main.rs` *was* the
crate, so the clap tree had exactly one consumer and no target outside the
binary could render anything from it. Now `src/lib.rs` owns the command tree,
the argument bootstrap and the dispatch behind `run()`, `src/main.rs` is three
lines over it, and `examples/gen_man.rs` renders the page from the same
`Cli::command()` the binary parses with. "Never drifts from the actual
commands" is unchanged; `clap_mangen` moves to dev-dependencies and leaves the
shipped binary. `h5i completion` stays: that one is genuinely run by a person.

A quiet gain from the split: the 36 unit tests in `main.rs` were invisible to
`cargo test --workspace --lib`, because a bin has no lib target. They run now.

The page also stops existing twice. It was mirrored to `docs/` on the pattern
`install.sh` uses, and that pattern does not transfer: `install.sh` is
duplicated only because it has to answer at exactly `/install.sh`, and a man
page can answer anywhere. So `man/` is deleted and `docs/man/man1/h5i.1` is
the only copy — `docs/` is published verbatim, so the file the repository has
is the file the site serves. The `man1/` subdirectory is kept, so
`MANPATH=$PWD/docs/man man h5i` still renders it. The CI sync gate goes away
with the second copy; the freshness gate simplifies to one path under `docs/`.

Regenerated: the man page (the diff is exactly the removed `H5I MAN` section)
and `docs/manual/index.html`, whose command table no longer lists `h5i man`.
