+++ README.md
@@ -227,7 +227,7 @@ receives a private, one-time copy of approved HOME state.
 - [Official Website](https://h5i.dev/): project overview, [Slides](https://h5i.dev/pitch/)
 - [MANUAL.md](MANUAL.md) / `man h5i`: full command reference
 - [CONTRIBUTING.md](CONTRIBUTING.md): we welcome contributions of any kind
-- `h5i man > ~/.local/share/man/man1/h5i.1`: install the man page (generated from the CLI)
+- `curl -fsSL https://h5i.dev/man/man1/h5i.1 -o ~/.local/share/man/man1/h5i.1`: install the man page (generated from the CLI)
 
 ---
