+++ docs/man/man1/h5i.1
@@ -40,9 +40,6 @@ Write or print the agent skill this binary carries
 h5i\-completion(1)
 Generate a shell completion script (bash, zsh, fish, ...); e.g. `h5i completion bash > /etc/bash_completion.d/h5i`
 .TP
-h5i\-man(1)
-Generate the roff man page from the CLI definition (so it never drifts from the actual commands); e.g. `h5i man > man/man1/h5i.1`
-.TP
 h5i\-help(1)
 Print this message or the help of the given subcommand(s)
 .SH VERSION
@@ -2356,17 +2353,5 @@ powershell
 .IP \(bu 2
 zsh
 .RE
-.SH "H5I MAN"
-Generate the roff man page from the CLI definition (so it never drifts from the actual commands); e.g. `h5i man > man/man1/h5i.1`
-.ie \n(.g .ds Aq \(aq
-.el .ds Aq '
-.SS SYNOPSIS
-\fBman\fR [\fB\-h\fR|\fB\-\-help\fR] 
-.ie \n(.g .ds Aq \(aq
-.el .ds Aq '
-.SS OPTIONS
-.TP
-\fB\-h\fR, \fB\-\-help\fR
-Print help
 .SH SEE ALSO
 Full narrative manual: \fBMANUAL.md\fR in the source tree, or the rendered \fB/manual/\fR page on the project site.
