+++ docs/manual/index.html
@@ -391,8 +391,8 @@ npx skills add h5i-dev/h5i  # same bytes, if you do not have the binary yet
 <td>Open a box someone else is sharing, from their ticket.</td>
 </tr>
 <tr>
-<td><code>h5i completion</code> / <code>h5i man</code></td>
-<td>Shell completions and the man page.</td>
+<td><code>h5i completion</code></td>
+<td>Shell completions for bash, zsh, fish and friends.</td>
 </tr>
 </tbody>
 </table>
